+++ src/transformers/generic.rs
@@ -57,27 +57,6 @@ pub fn get_generic_svc_url() -> Result<String> {
 pub fn prepare_generic_embedding_request(
     job_meta: VectorizeMeta,
     inputs: &[Inputs],
-) -> Result<EmbeddingRequest> {
-    let text_inputs = trim_inputs(inputs);
-    let payload = EmbeddingPayload {
-        input: text_inputs,
-        model: job_meta.transformer.to_string(),
-    };
-
-    let job_params: types::JobParams = serde_json::from_value(job_meta.params)?;
-
-    let svc_host = get_generic_svc_url().context("failed to get embedding service url from GUC")?;
-
-    Ok(EmbeddingRequest {
-        url: svc_host,
-        payload,
-        api_key: job_params.api_key,
-    })
-}
-
-pub fn prepare_generic_embedding_request_no_guc(
-    job_meta: VectorizeMeta,
-    inputs: &[Inputs],
     url: String,
 ) -> Result<EmbeddingRequest> {
     let text_inputs = trim_inputs(inputs);
